server/requirements.txt
erappdirs==1.4.3
asgiref==3.2.3
bcrypt==3.1.7
beautifulsoup4==4.8.2
bootstrap4==0.1.0
cachetools==4.1.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
click==7.1.1
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
distlib==0.3.0
dj-rest-auth==1.0.9
Django==2.2
django-allauth==0.42.0
django-bootstrap4==1.1.1
django-cors-headers==3.4.0
django-crispy-forms==1.8.1
django-rest-swagger==2.0.7
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
filelock==3.0.12
google-api-core==1.17.0
google-auth==1.14.1
google-cloud-core==1.3.0
google-cloud-storage==1.28.1
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.28.1
idna==2.9
itsdangerous==1.1.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
Pillow==7.1.2
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
simplejson==3.17.0
six==1.14.0
soupsieve==1.9.5
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.9
